everydayJAXRSstory v0.0.3
Mr. Hacker used to give tutions. He like to teach Kids, Math and Hindi. He charges per Hour basis to their parents. Mr. Hacker needs a web based service which can store the details of the hours tought by him to each Kid so that he can collect correct tution fees from parents of the students. Because Mr. Hacker is a geeky person so he wants the output of the services be in JSON or XML format. He wants the Name of the Kid, the List of Dates with the respective no. of hours tought to the Kid on that date as a detail in the output. Besause this application doesn't deal with huge data so use file system and java serialization to store and retrive the data.